Professor pages upload

+++ src/main/java/sda/studentmanagement/studentmanager/services/AttendanceServiceImplementation.java
@@ -8,9 +8,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sda.studentmanagement.studentmanager.domain.Attendance;
+import sda.studentmanagement.studentmanager.domain.Course;
 import sda.studentmanagement.studentmanager.domain.Session;
 import sda.studentmanagement.studentmanager.domain.User;
+import sda.studentmanagement.studentmanager.dto.AddAttendancesFormDto;
+import sda.studentmanagement.studentmanager.dto.EditAttendanceFormDto;
 import sda.studentmanagement.studentmanager.repositories.AttendanceRepository;
+import sda.studentmanagement.studentmanager.repositories.CourseRepository;
 import sda.studentmanagement.studentmanager.repositories.SessionRepository;
 import sda.studentmanagement.studentmanager.repositories.UserRepository;
 
@@ -18,6 +22,7 @@ import javax.persistence.EntityExistsException;
 import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +32,7 @@ public class AttendanceServiceImplementation implements AttendanceService {
     private final AttendanceRepository attendanceRepository;
     private final UserRepository userRepository;
     private final SessionRepository sessionRepository;
+    private final CourseRepository courseRepository;
 
     @Override
     public List<Attendance> getAttendances() {
@@ -88,6 +94,25 @@ public class AttendanceServiceImplementation implements AttendanceService {
     }
 
     @Override
+    public List<Attendance> getUserCourseAttendances(long userId, long courseId) {
+        Course course = courseRepository.findById(courseId);
+        User user = userRepository.findById(userId);
+        List<Attendance> userCourseAttendances = new ArrayList<>();
+
+        if (course != null && user != null) {
+            for(Attendance attendance : getAttendances()) {
+                if (attendance.getUser().equals(user) && attendance.getSession().getCourse().equals(course)) {
+                    userCourseAttendances.add(attendance);
+                }
+            }
+
+            return userCourseAttendances;
+        } else {
+            throw new EntityNotFoundException("Invalid Session, User ID");
+        }
+    }
+
+    @Override
     public Attendance saveAttendance(Attendance attendance) {
         Attendance findAttendance = attendanceRepository.findById(attendance.getId());
 
@@ -100,16 +125,69 @@ public class AttendanceServiceImplementation implements AttendanceService {
     }
 
     @Override
-    public Attendance editAttendance(Attendance attendance) {
-        // Todo
+    public void saveAttendances(AddAttendancesFormDto addAttendancesForm) {
+        Session session = sessionRepository.findById(addAttendancesForm.getSessionId());
+        List<Attendance> attendanceList = new ArrayList<>();
 
-        return null;
+        if (session == null) {
+            throw new EntityNotFoundException("Session not found");
+        } else if (addAttendancesForm.getAttendances().length == 0) {
+            throw new IllegalArgumentException("No Attendances marked");
+        } else {
+            for (int i = 0; i < addAttendancesForm.getAttendances().length; i++) {
+                Attendance attendance = new Attendance();
+                User user = userRepository.findById(addAttendancesForm.getAttendances()[i].getUserId());
+
+                if (user == null) {
+                    throw new EntityNotFoundException("User with ID: " + addAttendancesForm.getAttendances()[i].getUserId() + " not found");
+                }
+
+                Attendance existingAttendance = attendanceRepository.getExistingAttendanceByUserSessionId(user.getId(), session.getId());
+
+                if (existingAttendance != null) {
+                    existingAttendance.setDidAttend(addAttendancesForm.getAttendances()[i].isAttendance());
+
+                    attendanceRepository.save(existingAttendance);
+                } else {
+                    attendance.setSession(session);
+                    attendance.setUser(user);
+                    attendance.setDidAttend(addAttendancesForm.getAttendances()[i].isAttendance());
+
+                    attendanceList.add(attendance);
+                }
+            }
+
+            attendanceRepository.saveAll(attendanceList);
+        }
+    }
+
+    @Override
+    public Attendance editAttendance(long attendanceId, EditAttendanceFormDto attendanceForm) {
+        Attendance attendance = attendanceRepository.findById(attendanceId);
+
+        if (attendance != null) {
+            if (attendanceForm.getAttendance().isBlank()) {
+                throw new IllegalArgumentException("Please select Attendance");
+            }
+
+            attendance.setDidAttend(Objects.equals(attendanceForm.getAttendance(), "present"));
+
+            attendanceRepository.save(attendance);
+
+            return attendance;
+        } else {
+            throw new EntityNotFoundException("Session not found");
+        }
     }
 
     @Override
     public void deleteAttendance(long attendanceId) {
-        Attendance attendance = attendanceRepository.findById(attendanceId);
+        Attendance attendance = getAttendance(attendanceId);
 
-        attendanceRepository.delete(attendance);
+        if (attendance == null) {
+            throw new EntityNotFoundException("Attendance not found");
+        } else {
+            attendanceRepository.delete(attendance);
+        }
     }
 }
